Refatorado código da controller adicionado exceções e statusCode de retorno

+++ src/Controllers/PersonController.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Mvc;
 using src.Models;
 using src.Persistence;
 using Microsoft.EntityFrameworkCore;
+using System.Net;
 
 namespace src.Controllers;
 
@@ -16,36 +17,74 @@ public class PersonController : ControllerBase
     }   
 
     [HttpGet]
-    public List<Person> GetListPerson(){
-        // Person person = new Person("Henrique", 34, "12345678");
-        // Contract contract = new Contract("abc123", 50.0);
-        // person.contracts.Add(contract);
-        // person.contracts.Add(contract);
-        // person.contracts.Add(contract);
-        return _context.Persons.Include(p => p.contracts).ToList();
+    public ActionResult<List<Person>> GetListPerson(){
+        var result = _context.Persons.Include(p => p.contracts).ToList();
+        if (!result.Any()) {
+            return NoContent();
+        }
+        return Ok(result);
     }
 
     [HttpPost]
-    public Person Post([FromBody]Person person)
+    public ActionResult<Object> Post([FromBody]Person person)
     {
-        _context.Persons.Add(person);
-        _context.SaveChanges();
-        return person;
+        try
+        {
+            _context.Persons.Add(person);
+            _context.SaveChanges();     
+        }
+        catch (System.Exception)
+        {
+            return BadRequest();            
+        }
+        return Created("Criado", person);
     }
 
     [HttpPut("{id}")]
-    public string UpdatePerson([FromRoute]int id, [FromBody] Person person){
-        _context.Persons.Update(person);
-        _context.SaveChanges();
-        return "Dados do id " + id + " atualizados";
+    public ActionResult<Object> UpdatePerson([FromRoute]int id, [FromBody] Person person){
+
+        var result = _context.Persons.SingleOrDefault(e => e.id == id);
+        if (result is null){
+            return NotFound(new {
+                msg = "Registro não encntrado",
+                status = HttpStatusCode.BadRequest
+            });
+        }
+
+        try
+        {
+            _context.Persons.Update(person);
+            _context.SaveChanges();      
+        }
+        catch (System.Exception)
+        {
+            return BadRequest(new {
+                msg = $"Erros ao enviar solicitação de atualização do {id} da chamada",
+                status = HttpStatusCode.BadRequest
+            });
+        }
+        return Ok (new{
+            msg = $"Dados do id {id} atualizados",
+            status = HttpStatusCode.OK
+        });
     }
 
     [HttpDelete("{id}")]
-    public string Delete([FromRoute] int id){
+    public ActionResult<Object> Delete([FromRoute] int id){
         var result = _context.Persons.SingleOrDefault(e => e.id == id);
+        
+        if (result is null){
+            return BadRequest(new {
+                msg = "Conteudo inexistente, solicitação inválida",
+                status = HttpStatusCode.BadRequest
+            });
+        }
+
         _context.Persons.Remove(result);
         _context.SaveChanges();
-        return $"Foi deletado a pessoa do id {id}";
+        return Ok(new {
+           msg = $"Foi deletado a pessoa do id {id}",
+           status = HttpStatusCode.OK
+        });
     }
-
 }
